resolve issues
